fix worker depends_on
There was inconsistency in the worker compose yaml for depends_on. The worker requires both db and redis, and in the default docker-compose.yml it was missing db, causing worker to crash on `docker-compose up`. In two other compose files for stacks, where depends_on was used for other services, I added it for the worker to match.